requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
boto3==1.42.50
botocore==1.42.50
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
endee==0.1.13
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.13
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langsmith==0.7.3
msgpack==1.1.2
numpy==2.4.2
orjson==3.11.7
packaging==24.2
pillow==12.1.1
pinecone==8.0.1
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
websockets==15.0.1
whitenoise==6.11.0
xxhash==3.6.0
zstandard==0.25.0
